requirements.txt
bs4
beautifulsoup4
py-bing-search
lxml
unidecode
Jinja2
